requirements.txt
langgraph>=0.2.0
crewai>=0.80.0
openai>=1.0.0
playwright>=1.48.0
firecrawl>=1.0.0
httpx>=0.27.0
requests>=2.32.0
pyyaml>=6.0.2
rich>=13.9.0
loguru>=0.7.2
python-dotenv>=1.0.1
pydantic>=2.10.0
